Merged/fixed in changes from Markos serverless work
